+++ tailwind.config.ts
@@ -7,7 +7,8 @@ const config = {
     prefix: '',
     theme: {
         container: {
-            center: true,
+            //@ts-ignore temporary ignore as this can build successfully with it still.
+            center: 'true',
             padding: '2rem',
             screens: {
                 '2xl': '1400px'
@@ -59,23 +60,49 @@ const config = {
             },
             keyframes: {
                 'accordion-down': {
-                    from: { height: '0' },
-                    to: { height: 'var(--radix-accordion-content-height)' }
+                    from: {
+                        height: '0'
+                    },
+                    to: {
+                        height: 'var(--radix-accordion-content-height)'
+                    }
                 },
                 'accordion-up': {
-                    from: { height: 'var(--radix-accordion-content-height)' },
-                    to: { height: '0' }
+                    from: {
+                        height: 'var(--radix-accordion-content-height)'
+                    },
+                    to: {
+                        height: '0'
+                    }
                 },
                 'border-beam': {
                     '100%': {
                         'offset-distance': '100%'
+                    }
+                },
+                marquee: {
+                    from: {
+                        transform: 'translateX(0)'
+                    },
+                    to: {
+                        transform: 'translateX(calc(-100% - var(--gap)))'
+                    }
+                },
+                'marquee-vertical': {
+                    from: {
+                        transform: 'translateY(0)'
+                    },
+                    to: {
+                        transform: 'translateY(calc(-100% - var(--gap)))'
                     }
                 }
             },
             animation: {
                 'accordion-down': 'accordion-down 0.2s ease-out',
                 'accordion-up': 'accordion-up 0.2s ease-out',
-                'border-beam': 'border-beam calc(var(--duration)*1s) infinite linear'
+                'border-beam': 'border-beam calc(var(--duration)*1s) infinite linear',
+                marquee: 'marquee var(--duration) infinite linear',
+                'marquee-vertical': 'marquee-vertical var(--duration) linear infinite'
             }
         }
     },